feat: add ProcessDetector
